moved the events and constants to middleware
