+++ app/client/src/middlewares/webSocket/handleSocketEvent.ts
